+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/ChatApi.kt
@@ -178,6 +178,7 @@ internal interface ChatApi {
         messageId: String,
         set: Map<String, Any>,
         unset: List<String>,
+        skipEnrichUrl: Boolean = false,
     ): Call<Message>
 
     @CheckResult