+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/MessagesScreen.kt
@@ -98,6 +98,9 @@ import io.getstream.chat.android.ui.common.state.messages.list.SendAnyway
  * back button.
  * @param onHeaderTitleClick Handler for when the user taps on the header title section.
  * @param onChannelAvatarClick Handler called when the user taps on the channel avatar.
+ * @param skipPushNotification If new messages should skip triggering a push notification when sent. False by default.
+ * @param skipEnrichUrl If new messages being sent, or existing ones being updated should skip enriching the URL.
+ * If URL is not enriched, it will not be displayed as a link attachment. False by default.
  */
 @Suppress("LongMethod")
 @Composable
@@ -107,11 +110,19 @@ public fun MessagesScreen(
     onBackPressed: () -> Unit = {},
     onHeaderTitleClick: (channel: Channel) -> Unit = {},
     onChannelAvatarClick: () -> Unit = {},
+    skipPushNotification: Boolean = false,
+    skipEnrichUrl: Boolean = false,
 ) {
     val listViewModel = viewModel(MessageListViewModel::class.java, factory = viewModelFactory)
     val composerViewModel = viewModel(MessageComposerViewModel::class.java, factory = viewModelFactory)
     val attachmentsPickerViewModel =
         viewModel(AttachmentsPickerViewModel::class.java, factory = viewModelFactory)
+
+    val messageMode = listViewModel.messageMode
+
+    if (messageMode is MessageMode.MessageThread) {
+        composerViewModel.setMessageMode(messageMode)
+    }
 
     val backAction = {
         val isInThread = listViewModel.isInThread
@@ -136,7 +147,6 @@ public fun MessagesScreen(
             modifier = Modifier.fillMaxSize(),
             topBar = {
                 if (showHeader) {
-                    val messageMode = listViewModel.messageMode
                     val connectionState by listViewModel.connectionState.collectAsState()
                     val user by listViewModel.user.collectAsState()
 
@@ -166,7 +176,15 @@ public fun MessagesScreen(
                     onCancelAction = {
                         listViewModel.dismissAllMessageActions()
                         composerViewModel.dismissMessageActions()
-                    }
+                    },
+                    onSendMessage = { message ->
+                        composerViewModel.sendMessage(
+                            message.apply {
+                                this.skipPushNotification = skipPushNotification
+                                this.skipEnrichUrl = skipEnrichUrl
+                            }
+                        )
+                    },
                 )
             }
         ) {
@@ -189,12 +207,22 @@ public fun MessagesScreen(
                             val message = listViewModel.getMessageById(result.messageId)
 
                             if (message != null) {
-                                composerViewModel.performMessageAction(Reply(message))
+                                composerViewModel.performMessageAction(
+                                    Reply(
+                                        message.apply {
+                                            this.skipPushNotification = skipPushNotification
+                                            this.skipEnrichUrl = skipEnrichUrl
+                                        }
+                                    )
+                                )
                             }
                         }
 
                         MediaGalleryPreviewResultType.SHOW_IN_CHAT -> {
-                            listViewModel.scrollToMessage(result.messageId)
+                            listViewModel.scrollToMessage(
+                                messageId = result.messageId,
+                                parentMessageId = result.parentMessageId
+                            )
                         }
                         null -> Unit
                     }
@@ -204,7 +232,9 @@ public fun MessagesScreen(
 
         MessageMenus(
             listViewModel = listViewModel,
-            composerViewModel = composerViewModel
+            composerViewModel = composerViewModel,
+            skipPushNotification = skipPushNotification,
+            skipEnrichUrl = skipEnrichUrl,
         )
         AttachmentsPickerMenu(
             attachmentsPickerViewModel = attachmentsPickerViewModel,
@@ -212,7 +242,9 @@ public fun MessagesScreen(
         )
         MessageModerationDialog(
             listViewModel = listViewModel,
-            composerViewModel = composerViewModel
+            composerViewModel = composerViewModel,
+            skipPushNotification = skipPushNotification,
+            skipEnrichUrl = skipEnrichUrl,
         )
         MessageDialogs(listViewModel = listViewModel)
     }
@@ -224,11 +256,17 @@ public fun MessagesScreen(
  *
  * @param listViewModel The [MessageListViewModel] used to read state from.
  * @param composerViewModel The [MessageComposerViewModel] used to read state from.
+ * @param skipPushNotification If the message should skip triggering a push notification when sent. False by default. Note, only
+ * new messages trigger push notifications, updating edited messages does not.
+ * @param skipEnrichUrl If the message should skip enriching the URL. If URL is not enriched, it will not be
+ * displayed as a link attachment. False by default.
  */
 @Composable
 private fun BoxScope.MessageMenus(
     listViewModel: MessageListViewModel,
     composerViewModel: MessageComposerViewModel,
+    skipPushNotification: Boolean,
+    skipEnrichUrl: Boolean,
 ) {
     val selectedMessageState = listViewModel.currentMessagesState.selectedMessageState
 
@@ -238,7 +276,9 @@ private fun BoxScope.MessageMenus(
         listViewModel = listViewModel,
         composerViewModel = composerViewModel,
         selectedMessageState = selectedMessageState,
-        selectedMessage = selectedMessage
+        selectedMessage = selectedMessage,
+        skipPushNotification = skipPushNotification,
+        skipEnrichUrl = skipEnrichUrl,
     )
 
     MessagesScreenReactionsPicker(
@@ -246,6 +286,8 @@ private fun BoxScope.MessageMenus(
         composerViewModel = composerViewModel,
         selectedMessageState = selectedMessageState,
         selectedMessage = selectedMessage,
+        skipPushNotification = skipPushNotification,
+        skipEnrichUrl = skipEnrichUrl,
     )
 }
 
@@ -259,6 +301,10 @@ private fun BoxScope.MessageMenus(
  * perform actions.
  * @param selectedMessageState The state of the currently selected message.
  * @param selectedMessage The currently selected message.
+ * @param skipPushNotification If the message should skip triggering a push notification when sent. False by default. Note, only
+ * new messages trigger push notifications, updating edited messages does not.
+ * @param skipEnrichUrl If the message should skip enriching the URL. If URL is not enriched, it will not be
+ * displayed as a link attachment. False by default.
  */
 @Suppress("LongMethod")
 @OptIn(ExperimentalAnimationApi::class)
@@ -268,6 +314,8 @@ private fun BoxScope.MessagesScreenMenus(
     composerViewModel: MessageComposerViewModel,
     selectedMessageState: SelectedMessageState?,
     selectedMessage: Message,
+    skipPushNotification: Boolean,
+    skipEnrichUrl: Boolean,
 ) {
 
     val user by listViewModel.user.collectAsState()
@@ -313,6 +361,9 @@ private fun BoxScope.MessagesScreenMenus(
             message = selectedMessage,
             ownCapabilities = ownCapabilities,
             onMessageAction = { action ->
+                action.message.skipPushNotification = skipPushNotification
+                action.message.skipEnrichUrl = skipEnrichUrl
+
                 composerViewModel.performMessageAction(action)
                 listViewModel.performMessageAction(action)
             },
@@ -344,6 +395,9 @@ private fun BoxScope.MessagesScreenMenus(
             currentUser = user,
             message = selectedMessage,
             onMessageAction = { action ->
+                action.message.skipPushNotification = skipPushNotification
+                action.message.skipEnrichUrl = skipEnrichUrl
+
                 composerViewModel.performMessageAction(action)
                 listViewModel.performMessageAction(action)
             },
@@ -366,6 +420,10 @@ private fun BoxScope.MessagesScreenMenus(
  * perform actions.
  * @param selectedMessageState The state of the currently selected message.
  * @param selectedMessage The currently selected message.
+ * @param skipPushNotification If the message should skip triggering a push notification when sent. False by default. Note, only
+ * new messages trigger push notifications, updating edited messages does not.
+ * @param skipEnrichUrl If the message should skip enriching the URL. If URL is not enriched, it will not be
+ * displayed as a link attachment. False by default.
  */
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
@@ -374,6 +432,8 @@ private fun BoxScope.MessagesScreenReactionsPicker(
     composerViewModel: MessageComposerViewModel,
     selectedMessageState: SelectedMessageState?,
     selectedMessage: Message,
+    skipPushNotification: Boolean,
+    skipEnrichUrl: Boolean,
 ) {
 
     AnimatedVisibility(
@@ -398,6 +458,9 @@ private fun BoxScope.MessagesScreenReactionsPicker(
                 ),
             message = selectedMessage,
             onMessageAction = { action ->
+                action.message.skipPushNotification = skipPushNotification
+                action.message.skipEnrichUrl = skipEnrichUrl
+
                 composerViewModel.performMessageAction(action)
                 listViewModel.performMessageAction(action)
             },
@@ -468,6 +531,8 @@ private fun BoxScope.AttachmentsPickerMenu(
 private fun MessageModerationDialog(
     listViewModel: MessageListViewModel,
     composerViewModel: MessageComposerViewModel,
+    skipPushNotification: Boolean,
+    skipEnrichUrl: Boolean,
 ) {
     val selectedMessageState = listViewModel.currentMessagesState.selectedMessageState
 
@@ -485,7 +550,14 @@ private fun MessageModerationDialog(
                 when (action) {
                     DeleteMessage -> listViewModel.deleteMessage(message = message, true)
                     EditMessage -> composerViewModel.performMessageAction(Edit(message))
-                    SendAnyway -> listViewModel.performMessageAction(Resend(message))
+                    SendAnyway -> listViewModel.performMessageAction(
+                        Resend(
+                            message.apply {
+                                this.skipPushNotification = skipPushNotification
+                                this.skipEnrichUrl = skipEnrichUrl
+                            }
+                        )
+                    )
                     else -> {
                         // Custom events
                     }